Correct displaying of neighboring cells upon click
Actually this was how I had originally implemented it, but I thought it
was incorrect. Turns out that if you click a non-empty cell, it should
just display that cell and not the ones around it. Hmph. The way I had it
was actually more complicated but interesting too, but it made the game
too easy.